+++ packages/frontend/src/i18n/en.yaml
@@ -394,7 +394,7 @@ tips:
 
 # Tabulator Logs
 tabulator_logs:
-  # Note: tabulator_logs keys are the only ones that are stored in the database, so be careful when changing them
+  # Note: tabulator_logs keys are generated by the backend and they may be stored in the database in the future, so be extra careful when changing the keys
   star:
     # Score tiebreak
     # Note: this isn't considered a tiebreaker during the scoring round, but I'm keeping the naming consistent
